refactor(Hero): move styles
