+++ src/components/Hero/index.tsx
@@ -1,23 +1,6 @@
-import styled from '@emotion/styled'
 import { graphql, useStaticQuery } from 'gatsby'
-import BackgroundImage from 'gatsby-background-image'
 import React from 'react'
-
-const ImageBackground = styled(BackgroundImage)`
-  background-size: cover;
-  height: 100vh;
-  position: 'unset';
-  opacity: 0.95;
-  max-width: 100%;
-  padding: 0;
-  grid-row: 1/-1;
-  grid-column: 1/-1;
-  min-height: 100vh;
-
-  + * {
-    margin-top: 0;
-  }
-`
+import ImageBackground from './Hero.styles'
 
 const Hero = () => {
   const { image } = useStaticQuery(graphql`
